// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  SOAdapter: typeof SOAdapter;
  actions: typeof actions;
  auth: typeof auth;
  blog: typeof blog;
  chat: typeof chat;
  chatActions: typeof chatActions;
  cms: typeof cms;
  cmsActions: typeof cmsActions;
  contact: typeof contact;
  crons: typeof crons;
  extraCosts: typeof extraCosts;
  file: typeof file;
  generateblog: typeof generateblog;
  generateblogHelpers: typeof generateblogHelpers;
  helpers: typeof helpers;
  http: typeof http;
  invoice: typeof invoice;
  monthlysubscriptions: typeof monthlysubscriptions;
  monthlysubscriptionstest: typeof monthlysubscriptionstest;
  newsletter: typeof newsletter;
  organizations: typeof organizations;
  permissions: typeof permissions;
  project: typeof project;
  projectAgenda: typeof projectAgenda;
  quote: typeof quote;
  smartObjects: typeof smartObjects;
  workflow: typeof workflow;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;